requirements.txt
asgiref==3.2.5
Django==3.0.4
django-tinymce==2.8.0
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.1
